+++ SOSChildrenVillage/src/components/layout/Authenication/Register.jsx
@@ -1,6 +1,5 @@
 import React from "react";
-import LoginScreen from "../../components/layout/Authenication/LoginScreen";
-import RegisterScreen from "../../components/layout/Authenication/RegisterScreen";
+import RegisterScreen from "../../../pages/Authenication/RegisterScreen";
 import { Layout, Row, Col } from "antd"; // Ant Design components
 
 const { Content } = Layout;
@@ -10,11 +9,11 @@ export const Login = () => {
     <Layout>
       <Content style={{ padding: '50px 0' }}>
         <div className="container">
-          <div className="login__section--inner">
+          <div className="register__section--inner">
             <Row justify="center" align="middle">
               <Col xs={24} sm={24} md={12} lg={10}>
                 {/* LoginScreen Component */}
-                <LoginScreen />
+                <RegisterScreen />
               </Col>
             </Row>
           </div>
