+++ SOSChildrenVillage/src/components/layout/Booking/BookingHistoryPage.jsx
@@ -1,8 +1,8 @@
 import React from "react";
-import Header from "../../components/layout/HomePage/Header";
-import Nav from "../../components/layout/HomePage/Nav";
-import Footer from "../../components/layout/HomePage/Footer";
-import BookingHistory from "../../components/layout/Booking/BookingHistory";
+import Header from "../../../pages/Home/Header";
+import Nav from "../../../pages/Home/Nav";
+import Footer from "../../../pages/Home/Footer";
+import BookingHistory from "../../../pages/BookingPage/BookingHistory";
 
 const App = () => {
     return (
