+++ SOSChildrenVillage/src/pages/EventPage/EventDetail.jsx
