+++ SOSChildrenVillage/src/routes/imports.jsx
@@ -26,24 +26,24 @@ import TransactionManagement from '../pages/Admin/Transactions/TransactionManage
 import VillageManagement from '../pages/Admin/Villages/VillageManagement';
 
 // Home and Authentication imports
-import HomePages from '../pages/Home/HomePage';
-import Login from '../pages/Authenication/Login';
-import Register from '../pages/Authenication/Register';
-import UserDetail from '../pages/User/UserProfile';
-import EditUserProfile from '../pages/User/EditUserProfile';
-import ChangePasswordPage from '../pages/User/ChangePasswordPage';
+import HomePages from '../components/layout/HomePage/HomePage';
+import Login from '../components/layout/Authenication/Login';
+import Register from '../components/layout/Authenication/Register';
+import UserDetail from '../components/layout/User/UserProfile';
+import EditUserProfile from '../components/layout/User/EditUserProfile';
+import ChangePasswordPage from '../components/layout/User/ChangePasswordPage';
 
 // Event and Child Page imports
-import ChildDetailPage from '../pages/ChildPage/ChildDetailPage';
-import EventDetailPage from '../pages/EventPage/EventDetailPage';
+import ChildDetailPage from '../components/layout/Child/ChildDetailPage';
+import EventDetailPage from '../components/layout/Event/EventDetailPage';
 
 // Financial Page imports
-import Donate from '../pages/Donates/DonatePage';
-import DonateHistoryPage from '../pages/User/DonateHistoryPage';
-import PaymentReturnPage from '../pages/Donates/PaymentReturnPage';
-import VillageHistoryPage from '../pages/User/VillageHistoryPage';
-import ListHousePage from '../pages/BookingPage/ListHousePage';
-import BookingHistoryPage from '../pages/BookingPage/BookingHistoryPage';
+import Donate from '../components/layout/Donation/DonatePage';
+import DonateHistoryPage from '../components/layout/User/DonateHistoryPage';
+import PaymentReturnPage from '../components/layout/Donation/PaymentReturnPage';
+import VillageHistoryPage from '../components/layout/User/VillageHistoryPage';
+import ListHousePage from '../components/layout/Booking/ListHousePage';
+import BookingHistoryPage from '../components/layout/Booking/BookingHistoryPage';
 
 export {
     // Admin Layout
